Add files via upload
Created an Inventory management excel sheet which lets you keep the details of the Purchased/Sales of the items. It consists of some complex Excel formulas to make the job easier for the person working with it. It also shows the Stocks which are currently available. Please watch the full video for a detailed explanation of the Sheet created.  
(The figures used in the sheets are all hypothetical )