Add files via upload
I have created an excel sheet which has the data of different companies (sector wise divided) downloaded from google. I have assigned macros to it with a pivot table dividing the companies by sector and sub-sector wise and showing the change in percentage to it. Please check the video link for full explanation of me and my work